TwoStreamNetwork/requirements.txt
absl-py==0.9.0
adal==1.2.7
addict==2.4.0
antlr4-python3-runtime==4.8
appdirs==1.4.4
applicationinsights==0.11.10
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.2.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==19.3.0
awscli==1.22.13
backports-tempfile==1.0
backports-weakref==1.0.post1
beautifulsoup4==4.10.0
black==21.4b2
bleach==4.1.0
botocore==1.23.13
cachetools==4.0.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==8.0.1
colorama==0.4.3
contextlib2==0.6.0.post1
cryptography==3.4.7
cycler==0.10.0
databricks-cli==0.15.0
dataclasses==0.8
defusedxml==0.7.1
distro==1.6.0
docker==4.4.4
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.15.2
dotnetcore2==2.1.21
easydict==1.9
entrypoints==0.3
et-xmlfile==1.1.0
filelock==3.3.1
ftfy==6.0.3
fusepy==3.0.1
future==0.18.2
fvcore==0.1.3.post20210317
gast==0.3.3
gdown==4.4.0
gevent==21.12.0
gitdb==4.0.7
gitpython==3.1.18
google-auth==1.10.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
greenlet==1.1.2
grequests==0.6.0
grpcio==1.26.0
h5py==2.10.0
huggingface-hub==0.0.19
hydra-core==1.1.1
idna==2.8
imageio==2.13.5
imageio-ffmpeg==0.4.5
importlib-metadata==1.5.0
importlib-resources==5.4.0
iopath==0.1.9
ipywidgets==7.7.0
isodate==0.6.0
jeepney==0.7.1
jinja2==3.0.3
jmespath==0.10.0
joblib==0.14.1
jsonpickle==2.0.0
jsonschema==3.2.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras-applications==1.0.8
keras-preprocessing==1.1.0
keyboard==0.13.5
kiwisolver==1.1.0
lmdb==1.2.1
markdown==3.1.1
markupsafe==2.0.1
mistune==0.8.4
mlflow-skinny==1.19.0
mmcv==1.3.12
more-itertools==8.2.0
moviepy==1.0.3
msal==1.13.0
msal-extensions==0.2.2
msgpack==1.0.2
msrest==0.6.21
msrestazure==0.6.4
mypy-extensions==0.4.3
natsort==7.0.1
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
ndg-httpsclient==0.5.1
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.18.1
oauthlib==3.1.0
omegaconf==2.1.1
opencv-python==4.4.0.46
openpyxl==3.0.9
opt-einsum==3.1.0
packaging==21.0
pandas==1.0.3
pandocfilters==1.5.0
pathspec==0.9.0
pathtools==0.1.2
pillow==8.4.0
plotly==4.5.0
pluggy==0.13.1
portalocker==1.5.2
proglog==0.1.9
prometheus-client==0.13.1
promise==2.3
protobuf==3.19.4
psutil==5.9.0
py==1.8.1
pyarrow==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools
pycosat==0.6.3
pycparser==2.19
pydot==1.4.2
pygtrie==2.4.2
pyjwt==2.1.0
pyopenssl==19.1.0
pyparsing==2.4.6
pyrsistent==0.18.0
pysocks==1.7.1
pytest==5.4.0
python-dateutil==2.8.1
pytz==2019.3
pyyaml==5.3
regex==2021.10.8
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
ruamel-yaml==0.15.87
s3transfer==0.5.0
sacrebleu==1.4.4
sacremoses==0.0.46
scikit-learn==0.22.2.post1
scipy
secretstorage==3.3.1
send2trash==1.8.0
sentencepiece==0.1.90
sentry-sdk==1.5.10
setproctitle==1.2.3
shortuuid==1.0.8
simplejson==3.17.3
six==1.13.0
smmap==4.0.0
soupsieve==2.3.1
tabulate==0.8.9
tensorboard
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
timm==0.4.12
tokenizers==0.10.3
torch<2
torch-tb-profiler==0.2.1
torchtext==0.5.0
torchvision
tqdm==4.40.2
transformers==4.11.3
typed-ast==1.5.2
typing==3.7.4.1
typing-extensions==3.10.0.0
urllib3==1.25.7
wandb==0.12.14
warmup-scheduler==0.3
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==1.2.1
webvtt-py==0.4.6
werkzeug==0.16.0
widgetsnbextension==3.6.0
wrapt==1.11.2
xmltodict==0.12.0
yacs==0.1.8
yapf==0.31.0
zipp==3.1.0
zope-event==4.5.0
zope-interface==5.4.0
